On branch 15-build-point-management-service  Your branch is up to date with 'github/15-build-point-management-service'.  Changes to be committed: 	modified:   Device.py 	modified:   DeviceManagement.py 	modified:   PointManagement.py 	modified:   app.py
This commit closes Issue 12 Sort device properties in alphabetical order before committing to database.
Found that when sending readProperty requests for each of the Device Object properties available several devices reply back with an Abort: Segmentation-no-supported reply message for the object-list and active-subscriptions
properties. In this instance I will need to send readProperty multiple for an arbitrary number of object Id  numbers for each of the object types we're interested in which will generate a whole lot of traffic, and cannot be sent in a readPropertyMultiple because of this pesky segmentation-not-supported issue.